Backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # for PostgreSQL
sqlite3  # for SQLite (built-in)

# Cloud Services
boto3==1.34.0  # AWS S3
botocore==1.34.0
supabase==2.3.0  # Supabase client
pymongo==4.6.1  # MongoDB Atlas
motor==3.3.2  # Async MongoDB driver

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.0.0
PyJWT==2.8.0  # Additional JWT library
bcrypt==4.1.2

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# OCR and Document Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1

# Google Cloud Vision API - REMOVED
# google-cloud-vision==3.4.4
google-auth==2.23.4
google-cloud-storage==2.10.0

# NLP and Text Processing
spacy==3.7.2
transformers==4.36.0
nltk==3.8.1
textblob==0.17.1

# IoT and MQTT
paho-mqtt==1.6.1
redis==5.0.1

# Email Processing
fastapi-mail==1.4.1
aiosmtplib==3.0.1
imapclient==2.3.1
beautifulsoup4==4.12.2  # for HTML email parsing
email-validator==2.1.0

# File Processing
aiofiles==23.2.0
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Production
gunicorn==21.2.0
celery==5.3.4  # for background tasks
flower==2.0.1  # celery monitoring

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.19.0
structlog==23.2.0

# CORS
python-cors==1.0.0